+++ lib/routes/matataki/ipfs/minetoken/related.js
@@ -1,4 +1,4 @@
-const matatakiUtils = require('@/utils/matataki-utils');
+const matatakiUtils = require('@/routes/matataki/utils/matataki-utils');
 
 module.exports = async (ctx) => {
