+++ lib/routes/matataki/posts/scoreranking.js
@@ -1,8 +1,10 @@
-const matatakiUtils = require('@/utils/matataki-utils');
+const matatakiUtils = require('@/routes/matataki/utils/matataki-utils');
 
 module.exports = async (ctx) => {
 
-    const items = await matatakiUtils.getPostFeedItems(`https://api.smartsignature.io/posts/scoreRanking`);
+    const querystring = ctx.request.querystring;
+
+    const items = await matatakiUtils.getPostFeedItems(`https://api.smartsignature.io/posts/scoreRanking?${querystring}`);
 
     ctx.state.data = {
         title: `瞬Matataki - 热门作品`,
