+++ lib/routes/matataki/utils/matataki-utils.js
@@ -10,6 +10,7 @@ const MATATAKI_WEB_URL = 'https://www.matataki.io/';
  */
 const IPFS_GATEWAY_URL = 'https://ipfs.io';
 
+
 /**
  * 以`get` 方式调用Matataki API的简单封装
  *
